Added a little map to the index page. Based on the center of the map, displays the questions (based on Dennis' backend code) for which the center is within the radius. Automatically updates when the map center changes.
Needs to be updated once Dennis updates his code
to get all the questions for a bounding box, not
just the center.

Right now, just displays all possible questions
at the bottom of the page.

ABOUT HEROKU - Added filter to application.js but
does not work on Heroku.

Removed filter from head and added it to application.js.